refactor youtube utils and format code with prettier
