+++ server/tsconfig.json
@@ -4,9 +4,7 @@
     "allowJs": true,
     "strict": true,
     "esModuleInterop": true,
-    "module": "commonjs",
+    "module": "commonjs"
   },
-  "include": [
-    "../global.d.ts"
-  ]
+  "include": ["../global.d.ts"]
 }
